Add a BaseScriptableEvent class with no args for better editor support (should fix Odin issues)
