--- Packages/com.chark.scriptable-events/Editor/FallbackScriptableEventEditor.cs
@@ -1,9 +1,0 @@
-﻿using UnityEditor;
-
-namespace ScriptableEvents.Editor
-{
-    [CustomEditor(typeof(BaseScriptableEvent<>), true)]
-    internal class FallbackScriptableEventEditor : BaseScriptableEventEditor
-    {
-    }
-}
