--- Packages/com.chark.scriptable-events/Editor/FallbackScriptableEventEditor.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 8638f4b923314588bf4d37bc888d5a08
-timeCreated: 1623304352
